fm_radio/fm_basics.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test for Power API</title>
</head>
<body>
<script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>

<script type="application/javascript">

function postMsg(msg)
{
  parent.postMessage(msg, "http://mochi.test:8888");
}

function ok(value, msg) {
  var result = {
   "type": "ok",
   "result": value ? true : false, 
   "msg": msg  
  };
  postMsg(result);
}

ok(navigator.mozFMRadio, "mozFMRadio interface is enabled");

mozFMRadio = navigator.mozFMRadio; 


/** Test permission **/

parent.postMessage("finished", "*");

</script>
</pre>
</body>
</html>
